+++ HISTORY.md
@@ -9,6 +9,7 @@ History
   * `CompCtorP`  → `ReactComponentC.ReqProps`
   * `CompCtorOP` → `ReactComponentC.DefaultProps`
   * `CompCtorNP` → `ReactComponentC.ConstProps`
+* Added `Simulation` for composition and abstraction of `ReactTestUtils.Simulate` procedures.
 
 ### 0.4.1 ([commit log](https://github.com/japgolly/scalajs-react/compare/v0.4.0...v0.4.1))